requirements.txt
aiofiles==23.2.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cryptography==45.0.3
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
docx2txt==0.9
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fonttools==4.58.1
fpdf==1.7.2
fpdf2==2.8.3
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.2.2
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keybert==0.9.0
kiwisolver==1.4.7
langchain==0.3.25
langchain-core==0.3.59
langchain-text-splitters==0.3.8
langsmith==0.3.42
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.38.2
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
openai==1.78.0
openpyxl==3.1.5
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==10.4.0
protobuf==3.20.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
Pympler==1.1
PyMuPDF==1.26.0
pyodbc==5.2.0
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.13
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.33.0
semantic-version==2.10.0
semver==3.0.4
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
streamlit==1.45.0
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.12.0
toolz==1.0.0
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.3
validators==0.35.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wsproto==1.2.0
XlsxWriter==3.2.3
zipp==3.21.0
zstandard==0.23.0
